+++ domain/src/main/java/org/akvo/flow/domain/interactor/forms/RegistrationFormResult.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 Stichting Akvo (Akvo Foundation)
+ * Copyright (C) 2021 Stichting Akvo (Akvo Foundation)
  *
  * This file is part of Akvo Flow.
  *
@@ -17,18 +17,8 @@
  * along with Akvo Flow.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.akvo.flow.presentation.record
+package org.akvo.flow.domain.interactor.forms
 
-import org.akvo.flow.domain.entity.User
+import org.akvo.flow.domain.entity.DomainForm
 
-interface RecordView {
-    fun showBootStrapPendingError()
-    fun showMissingCascadeError()
-    fun navigateToForm(formId: String, formInstanceId: Long)
-    fun navigateToForm(formId: String, user: User)
-    fun showDataPointTitle(displayName: String)
-    fun showDataPointError()
-    fun showMissingUserError()
-    fun showFormNotFound()
-    fun displayWarningDialog(formName: String, formId: String, user: User)
-}
+data class RegistrationFormResult(val form: DomainForm?)