+++ domain/src/main/java/org/akvo/flow/domain/repository/ApkRepository.java
@@ -29,5 +29,5 @@ public interface ApkRepository {
 
     Observable<Boolean> saveApkDataPreference(@NonNull ApkData apkData);
 
-    Observable<ApkData> getApkDataPreference();
+    ApkData getApkDataPreference();
 }